pyproject.toml
[project]
name = "ch4net"
version = "0.1.0"
description = "U-Net for methane plume segmentation from Sentinel-2 imagery"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "torchvision",
    "numpy",
    "scikit-learn",
    "tqdm",
    "matplotlib",
    "pystac-client>=0.9.0",
    "rioxarray>=0.19.0",
    "rasterio>=1.4.4",
    "shapely>=2.1.2",
    "requests>=2.32.5",
    "scipy>=1.15.3",
    "huggingface-hub>=1.4.1",
    "pillow>=12.1.1",
]

[dependency-groups]
dev = [
    "modal>=1.3.3",
]
